Apply Bandpass Filter to Fluorescence Intensities

Apply a bandpass like filter to the fluorescence intensities within a single gate.

Usage

applyBandpass(fdmat, bandpass, gate = 1)

Arguments

fdmat

An object of class-fdmat as produced by makefdmat.

bandpass

Numeric vector length two, holding the lower and upper boundaries of the bandpass filter.

gate

Numeric or character length one, defining the gate on which the bandpass should be applied. Defaults to 1. (See 'fdmat@metadata' for possible values for gate name resp. number.)

Value

The same object of class 'fdmat' as was provided to the argument 'fdmat', but with modified range of fluorescence intensities in the gate specified under 'gate'.

Examples

Please refer to https://bpollner.github.io/flowdex/ for executable examples in the tutorial dataset.

See also