Manually Update the Settings

Source the settings-list as defined in the settings.R file and create the object 'stn' in its environment. The settings-file system is based on functionality from package uniset.

Usage

fd_update_settings(silent = FALSE)

Arguments

silent

Logical

Value

A list holding key-value pairs as defined in the 'flowdex_settings.R' file.

See also

Examples

assign("get_settings_from_flowdex_package_root", TRUE, pos=.GlobalEnv)
# only required to make the examples run automatically
# you should not call 'assign' if you run the examples manually
# the effect of setting 'get_settings_from_flowdex_package_root' to TRUE
# is that the file 'flowdex_settings.R' in 'root' of the installed package
# 'flowdex' will be sourced instead of the one in the user-defined location.
#
stn <- fd_update_settings()
str(stn)
#> List of 51
#>  $ gen_autoUpdateSettings: logi TRUE
#>  $ dV_verbose            : logi TRUE
#>  $ dV_use_volumeData     : logi TRUE
#>  $ foN_gating            : chr "gating"
#>  $ foN_fcsFiles          : chr "fcsFiles"
#>  $ foN_rawData           : chr "rawdata"
#>  $ foN_templates         : chr "templates"
#>  $ foN_dictionary        : chr "dictionary"
#>  $ foN_plots             : chr "plots"
#>  $ fiN_gate              : chr "polyGate"
#>  $ fiN_gateStrat         : chr "gateStrat"
#>  $ dE_exportType         : chr "xlsx"
#>  $ dE_exportGate         : NULL
#>  $ fiN_dataExport        : chr "flscData"
#>  $ dV_charEventsPerVolume: chr "evpv"
#>  $ dV_tx                 : chr "fjbiexp"
#>  $ dV_channel            : chr "A$"
#>  $ dV_comp               : logi FALSE
#>  $ dV_gateStratInputType : chr "xlsx"
#>  $ dV_dictionaryType     : chr "xlsx"
#>  $ dV_channelsForPGG     : chr [1:2] "FITC.A" "PerCP.A"
#>  $ dV_channelBoundaries  : num [1:4] 1250 4000 0 4000
#>  $ dV_doRecalcToVolume   : logi TRUE
#>  $ dV_volumeFactor       : num 1e+06
#>  $ dV_volumeUnit         : chr "ml"
#>  $ dV_resolution         : num 220
#>  $ dV_cutoff_raw         : num 10
#>  $ dV_cutoff_Vol         : num 125
#>  $ dV_cutoff_apply       : logi TRUE
#>  $ dV_ignoreEdgePercent  : logi FALSE
#>  $ dV_doSmooth           : logi TRUE
#>  $ dV_smooth_n           : num 11
#>  $ dV_smooth_p           : num 5
#>  $ dV_charBeforeFlscNr   : chr "flsc"
#>  $ dD_useDictionary      : logi TRUE
#>  $ dD_dict_name          : chr "dictionary"
#>  $ dD_dict_type          : chr "xlsx"
#>  $ dD_elementSep         : chr ":"
#>  $ dD_groupSep           : chr ";"
#>  $ dD_classVarPrefix     : chr "C_"
#>  $ dD_numericVarPrefix   : chr "Y_"
#>  $ dG_locatorLine        : chr "darkgreen"
#>  $ dG_locatorLineWidth   : num 1.4
#>  $ dG_gateShowColor      : chr "violet"
#>  $ dG_pdf_height         : num 10
#>  $ dG_pdf_width          : num 14
#>  $ dG_nrBins             : num 128
#>  $ dG_nonFluorescenceChar: chr [1:2] "FSC.A" "SSC.A"
#>  $ dG_alphaForLegends    : num 200
#>  $ dG_RcolorBrewerPal    : chr "Dark2"
#>  $ last                  : num 0