Plot Gates on all flowframes in a gating set

Plot all available gates on all flowframes in a gating set and add layers for the number of events. (In raw format, i.e. *not* re-calculated to volume!)

Usage

plotgates(
  gs,
  ti = "",
  spl = NULL,
  fns = NULL,
  plotAll = FALSE,
  toPdf = TRUE,
  x = NULL,
  y = NULL,
  name.dict = ".",
  foN.dict = ".",
  type.dict = ".",
  foN.plots = "."
)

Arguments

gs

A gating set.

ti

Character length one, a possible character added to the title of the gate-plot.

spl

Character length one. The name of the column in the cyTags that should be used to split by before plotting. If left at the default NULL, no splitting is performed. Possible values for 'spl' are the column names of the cyTags saved in the object of class 'fdmat' as produced by makefdmat.

fns

Character length one. The filename suffix of the possible pdf.

plotAll

Logical. If left at the default FALSE, only the gates where the parameter keepData in the gating strategy is set to TRUE are plotted. If set to TRUE, all gates within the gating strategy file will be plotted.

toPdf

Logical. If the plots should be saved in a pdf. Defaults to TRUE

x

Character length one. The name of channel where data was acquired to be displayed on the x-axis. Only applies if a gating set without applied gate is provided to the argument gs.

y

Character length one. The name of channel where data was acquired to be displayed on the y-axis. Only applies if a gating set without applied gate is provided to the argument gs.

name.dict

Character length one. The name of the dictionary. If left at the default '.', the value as defined in the settings file (key 'dD_dict_name') will be used.

foN.dict

Character length one. The name of the folder where the dictionary resides. If left at the default '.', the value as defined in the settings file (key 'foN_dictionary') will be used.

type.dict

Character length one. The filetype of the dictionary. Can be one of 'csv' or 'xlsx'. If left at the default '.', the value as defined in the settings file (key 'dD_dict_type') will be used.

foN.plots

Character length one. The name of the folder where possible PDFs should be saved in. If left at the default '.', the value as defined in the settings file (key 'foN_plots') will be used.

Details

Plotting is performed by the function ggcyto. If a gating set without applied gates is provided to the first argument, parameters plotAll, spl and toPdf do not apply.

See also

Examples

if (FALSE) {
gs <- makeAddGatingSet()
plotgates(gs)
#
fdm <- makefdmat(gs)
fdm@cyTags # look at column names for splitting
plotgates(gs, spl="fooBar") # to split the plot by values contained in column 
# 'fooBar'.
}